fix: change TelegramChatId type from int to long in UserDto

+++ Catebi.Api.Domain/Features/AdoptionBot/ViewModels/UserDto.cs
@@ -6,9 +6,9 @@ public class UserDto
     public string? RecordId { get; set; }
     public string Name { get; set; }
     public string Telegram { get; set; }
-    public int TelegramChatId { get; set; }
+    public long TelegramChatId { get; set; }
     public bool? UsePayedAccount { get; set; }
     public bool? IsVolunteer { get; set; }
     public string? Status { get; set; }
     public string? Role { get; set; }
-} 
+}
